% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDP.R
\name{NDP}
\alias{NDP}
\title{Launch 'NDP' Interface}
\usage{
NDP()
}
\value{
Nothing
}
\description{
NDP() loads interactive user interface built using R 'shiny'.
}
\details{
Start NDP

The interactive user interface is to provide an easy way for people who are learning normal distribution. Includes example data for testing out a few example analysis.
}
\examples{
if(interactive()){
library(rmarkdown)
NDP()
}
}
\keyword{NDP}
